feat: salvando os dados da umm no hive

+++ lib/app/modules/internal_modules/user/data/models/user_umm_model.dart
@@ -69,7 +69,7 @@ class Grad {
 class Matriculas {
 
   @HiveField(0)
-  BigInt? id;
+  int? id;
 
   @HiveField(1)
   String? matricula;
